rename: properties 파일 이동
