+++ src/main/java/codesquad/fineants/global/init/SetupDataLoader.java
@@ -30,6 +30,10 @@ import codesquad.fineants.domain.stock.repository.StockRepository;
 import codesquad.fineants.global.errors.errorcode.MemberErrorCode;
 import codesquad.fineants.global.errors.errorcode.RoleErrorCode;
 import codesquad.fineants.global.errors.exception.FineAntsException;
+import codesquad.fineants.global.init.properties.AdminProperties;
+import codesquad.fineants.global.init.properties.ManagerProperties;
+import codesquad.fineants.global.init.properties.RoleProperties;
+import codesquad.fineants.global.init.properties.UserProperties;
 import codesquad.fineants.global.security.oauth.dto.MemberAuthentication;
 import codesquad.fineants.infra.s3.service.AmazonS3DividendService;
 import codesquad.fineants.infra.s3.service.AmazonS3StockService;
