+++ src/main/java/codesquad/fineants/global/init/properties/AdminProperties.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.global.init;
+package codesquad.fineants.global.init.properties;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.bind.ConstructorBinding;
