+++ src/main/java/codesquad/fineants/global/init/properties/RoleProperties.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.global.init;
+package codesquad.fineants.global.init.properties;
 
 import java.util.List;
