+++ src/test/java/codesquad/fineants/global/init/SetupDataLoaderTest.java
@@ -31,6 +31,9 @@ import codesquad.fineants.domain.stock.domain.dto.response.StockDataResponse;
 import codesquad.fineants.domain.stock.domain.entity.Stock;
 import codesquad.fineants.domain.stock.repository.StockRepository;
 import codesquad.fineants.domain.stock.service.StockCsvReader;
+import codesquad.fineants.global.init.properties.AdminProperties;
+import codesquad.fineants.global.init.properties.ManagerProperties;
+import codesquad.fineants.global.init.properties.UserProperties;
 import codesquad.fineants.global.security.oauth.dto.MemberAuthentication;
 import codesquad.fineants.infra.s3.service.AmazonS3DividendService;
 import codesquad.fineants.infra.s3.service.AmazonS3StockService;
